requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
python-telegram-bot==20.7
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0